requirements.txt
requests
click==7.1.1
Flask==1.1.2
PyYAML==5.3.1
Werkzeug==1.0.1